+++ FR3DLdapBundle.php
@@ -4,6 +4,7 @@ namespace FR3D\LdapBundle;
 
 use FR3D\LdapBundle\Security\Factory\FormLoginLdapFactory;
 use FR3D\LdapBundle\Security\Factory\HttpBasicLdapFactory;
+use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -20,6 +21,7 @@ class FR3DLdapBundle extends Bundle
     {
         parent::build($container);
 
+        /** @var SecurityExtension $extension */
         $extension = $container->getExtension('security');
         $extension->addSecurityListenerFactory(new FormLoginLdapFactory());
         $extension->addSecurityListenerFactory(new HttpBasicLdapFactory());