+++ Tests/Driver/ZendLdapDriverMockedTest.php
@@ -3,7 +3,8 @@
 namespace FR3D\LdapBundle\Tests\Driver;
 
 use FR3D\LdapBundle\Driver\ZendLdapDriver;
-use FR3D\LdapBundle\Tests\TestUser;
+use FR3D\LdapBundle\Model\LdapUser;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Zend\Ldap\Exception\LdapException as ZendLdapException;
 use Zend\Ldap\Ldap;
 
@@ -13,7 +14,7 @@ use Zend\Ldap\Ldap;
 class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var \Zend\Ldap\Ldap
+     * @var \Zend\Ldap\Ldap|\PHPUnit_Framework_MockObject_MockObject
      */
     protected $zend;
 
@@ -64,7 +65,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'test_username';
         $password = 'password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $this->zend->expects($this->once())
@@ -79,7 +80,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'test_username';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $this->zend->expects($this->once())
@@ -94,7 +95,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'bad_username';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $this->zend->expects($this->once())
@@ -109,7 +110,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $dn       = 'uid=test_username,ou=example,dc=com';
         $password = 'password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setDn($dn);
 
         $this->zend->expects($this->once())
@@ -124,7 +125,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $dn       = 'uid=test_username,ou=example,dc=com';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setDn($dn);
 
         $this->zend->expects($this->once())
@@ -139,6 +140,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'username';
         $password = 'password';
+        /** @var UserInterface|\PHPUnit_Framework_MockObject_MockObject $user */
         $user     = $this->getMock('Symfony\Component\Security\Core\User\UserInterface');
 
         $user->expects($this->once())