+++ Tests/Driver/ZendLdapDriverTest.php
@@ -3,7 +3,8 @@
 namespace FR3D\LdapBundle\Tests\Driver;
 
 use FR3D\LdapBundle\Driver\ZendLdapDriver;
-use FR3D\LdapBundle\Tests\TestUser;
+use FR3D\LdapBundle\Model\LdapUser;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Zend\Ldap\Ldap;
 
 /**
@@ -62,12 +63,16 @@ class ZendLdapDriverTest extends AbstractLdapDriverTest
     // Bind (bindRequireDn=false)
     /**
      * @dataProvider provideTestBind
+     *
+     * @param string $bind_rdn
+     * @param string $password
+     * @param bool $expect
      */
     public function testBind($bind_rdn, $password, $expect)
     {
         global $ldapServer;
 
-        $user = new TestUser();
+        $user = new LdapUser();
         $user->setUsername($bind_rdn);
 
         $ldapServer->expects($this->once())
@@ -98,6 +103,7 @@ class ZendLdapDriverTest extends AbstractLdapDriverTest
 
         $username = 'username';
         $password = 'password';
+        /** @var UserInterface|\PHPUnit_Framework_MockObject_MockObject $user */
         $user     = $this->getMock('Symfony\Component\Security\Core\User\UserInterface');
 
         $user->expects($this->once())