+++ Tests/Security/Authentication/LdapAuthenticationProviderTest.php
@@ -2,11 +2,12 @@
 
 namespace FR3D\LdapBundle\Tests\Security\Authentication;
 
+use FR3D\LdapBundle\Model\LdapUser;
 use FR3D\LdapBundle\Security\Authentication\LdapAuthenticationProvider;
-use FR3D\LdapBundle\Tests\TestUser;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Symfony\Component\Security\Core\Exception\AuthenticationServiceException;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\User\UserCheckerInterface;
 
 class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
 {
@@ -31,6 +32,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
+        /** @var UserCheckerInterface|\PHPUnit_Framework_MockObject_MockObject $userChecker */
         $userChecker                = $this->getMock('Symfony\Component\Security\Core\User\UserCheckerInterface');
         $providerKey                = 'provider_key';
         $this->userProvider         = $this->getMock('Symfony\Component\Security\Core\User\UserProviderInterface');
@@ -44,7 +46,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
     {
         $method   = $this->setMethodAccessible('retrieveUser');
         $username = 'test_username';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $token    = new UsernamePasswordToken($username, 'password', 'provider_key', array());
 
         $this->userProvider->expects($this->once())
@@ -115,7 +117,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method   = $this->setMethodAccessible('checkAuthentication');
         $username = 'username';
         $password = 'password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $token    = new UsernamePasswordToken($username, $password, 'provider_key', array());
@@ -136,7 +138,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method = $this->setMethodAccessible('checkAuthentication');
         $username = 'username';
         $password = 'password';
-        $user = new TestUser();
+        $user = new LdapUser();
         $user->setUsername($username);
 
         $token = new UsernamePasswordToken($user, $password, 'provider_key', array());
@@ -162,7 +164,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method   = $this->setMethodAccessible('checkAuthentication');
         $username = 'bad_username';
         $password = 'other_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $token    = new UsernamePasswordToken($username, $password, 'provider_key', array());
@@ -181,7 +183,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method   = $this->setMethodAccessible('checkAuthentication');
         $username = 'test_username';
         $password = 'password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
         $token    = new UsernamePasswordToken($username, $password, 'provider_key', array());
 
@@ -204,7 +206,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method   = $this->setMethodAccessible('checkAuthentication');
         $username = 'test_username';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
         $token    = new UsernamePasswordToken($username, $password, 'provider_key', array());
 
@@ -225,7 +227,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method   = $this->setMethodAccessible('checkAuthentication');
         $username = 'test_username';
         $password = '';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $token = new UsernamePasswordToken($username, $password, 'provider_key', array());
@@ -238,7 +240,7 @@ class LdapAuthenticationProviderTest extends \PHPUnit_Framework_TestCase
         $method   = $this->setMethodAccessible('checkAuthentication');
         $username = 'test_username';
         $password = '0';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $token = new UsernamePasswordToken($username, $password, 'provider_key', array());